Don't use invalid triangulated points
Summary:
pygeometry.triangulate_bearings_midpoint returns a valid flag and a possibly invalid triangulated point.

At some point in the past we expected the point to be None if it was invalid. This is no longer true. I imagine it changed during the C++ redo. Now the boolean flag returned from c++ has to be explicitly checked.

This broke object merge and also the triangulation of GCPs, where this was also not being checked.

I've also removed duplicate functionality to triangulate a gcp that I found while looking for calls of pygeometry.triangulate_bearings_midpoint and renamed usages such that the meaning of the output of this function is clearer.

Reviewed By: paulinus

Differential Revision: D26368979

fbshipit-source-id: 8effa919d1112060d38b024f74da288b7f206dea