Fix LRU cache issues
Summary:
# Context
Some feature related code parts rely on the `lru_cache`. However, the cache was not freed before and this introduced cyclic reference, preventing Python's `gc`  from freeing objects correctly.

This **only** became visible as a problem when doing something when an object is destroyed, e.g. storing files in the `__del__` method.

# Fix:
- Call `clear_cache()` method for `FeatureLoader` (was not called before!)
- Fixed missing brackets in one of the `clear_cache` calls
- Implemented `clear_cache()` for `vlad`

Reviewed By: YanNoun

Differential Revision: D29910299

fbshipit-source-id: 07277353d7d4a3289bd64230b51d5b1bbbfb83ba