feat: set orientation to auto in OpenSfM
Summary:
This Diff set the orientation method au the auto mode for OpenSfM dataset.

In auto mode, OpenSfM checks for the conditionning of the GPS/GCP constraints and automatically switch between `naive` and `orientation_prior`.

Before it was set to `horizontal` `orientation_prior`, which caused issue in vertically oriented datasets (aerial).

Reviewed By: paulinus

Differential Revision: D30929452

fbshipit-source-id: f813ff5047c3a3cfde29f42cc3b086b00bf0c302