feat: rigs as first-class citizen
Summary:
This Diff makes OpenSfM relying on rigs to represent all the shots. When creating a shot :
 - A rig instance for the shot will be created
 - The corresponding rig camera will be a rig camera automatically created from any camera and initialized with identity pose. This open the door to modeling leverarm shift between a single camera and its GPS receiver.
 - When "copying" a shot to a new reconstruction, corresponding `rig_camera` and `rig_instance` are also created.

OpenSfM still support the use of legacy "single-shots", however, any old reconstruction loaded will be created in-memory with all-shots-are-rig : the change is backward compatible and transparent.

More-over, it is still possible to change the pose of shot, given its `RigCamera` has identity transform. Of-course, this isn't allowed if it's not the case and setting the extrinsics should be done through the `RigInstance` and the `RigCamera` pose (normal rig workflow).

For now, rigs are still not modelled during the `similarity_averaging`, but the switch will be done in the next Diff.

Reviewed By: fabianschenk

Differential Revision: D30607361

fbshipit-source-id: 29db16edbff6a5330e2426c04848e3b2086139e3