fix map() in python3 (#906)
Summary:
In python3, map() function returns a map object which acts like an iterator, not a list. So, in order to append a map to a list, first we should convert that map into a list, then append the new list. Or else, append() will raise an error.

Pull Request resolved: https://github.com/mapillary/OpenSfM/pull/906

Reviewed By: YanNoun, paulinus

Differential Revision: D37064348

Pulled By: fabianschenk

fbshipit-source-id: 149244242bff47d3c1a00b58d0459b8d63a940aa