max_angle argument for object triangulation (#909)
Summary:
Pull Request resolved: https://github.com/mapillary/OpenSfM/pull/909

Triangulation should have both a minimum and a maximum angle (180degrees is also not good for triangulation). This adds the max threshold.

While a good default for max_angle might be 180-min_angle, this is left as a parameter. If the keypoint location is unstable for large viewpoint changes, a max angle closer to 90 degrees or even smaller might be desirable.

Reviewed By: fabianschenk

Differential Revision: D36241730

fbshipit-source-id: 30c7b7e122ea78776a435465c239f0167ef5595c